Restructuring of code
