+++ src/components/bytes/View/ByteStepper.tsx
@@ -1,7 +1,7 @@
+import ByteStepperItemWithProgressBar from '@/components/bytes/View/ByteStepperItem/ByteStepperItemWithProgressBar';
 import { UseGenericViewByteHelper } from '@/components/bytes/View/useGenericViewByte';
 import { ByteDetailsFragment, ByteStepFragment, ProjectByteFragment, SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
 import { useMemo } from 'react';
-import WithCarouselAndProgress1 from './NewByteStyles/style1/WithCarouselAndProgress1';
 
 type Props = {
   viewByteHelper: UseGenericViewByteHelper;
@@ -14,7 +14,7 @@ function ByteViewStepper({ viewByteHelper, byte, space }: Props) {
     return byte.steps?.[viewByteHelper.activeStepOrder] || byte.steps[0];
   }, [viewByteHelper.activeStepOrder, byte.steps]);
 
-  return <WithCarouselAndProgress1 viewByteHelper={viewByteHelper} byte={byte} step={activeStep} space={space} />;
+  return <ByteStepperItemWithProgressBar viewByteHelper={viewByteHelper} byte={byte} step={activeStep} space={space} />;
 }
 
 export default ByteViewStepper;
