+++ src/components/bytes/View/ByteStepperItem/Progress/ContinuousStepIndicatorProgress.module.scss
