+++ src/components/bytes/View/ByteStepperItem/Progress/ContinuousStepIndicatorProgress.tsx
