+++ src/components/bytes/View/ByteStepperItem/Progress/StepIndicatorProgress.module.scss
