+++ src/components/bytes/View/ByteStepperItem/Progress/StepIndicatorProgress.tsx
