--- src/components/bytes/View/NewByteStyles/style2/WithCarouselAndProgress2.module.scss
@@ -1,12 +1,0 @@
-.stepContent {
-  @media (min-width: 976px) {
-    .previous-text {
-      display: none;
-    }
-    border-radius: 0.5rem;
-  }
-}
-
-.styledH1 {
-  color: var(--heading-color);
-}
